Code refactor eventlist for DL2
