+++ dl2/eventlist_v4.py
@@ -80,16 +80,14 @@ class GFhandler2:
 
         #self.logger.debug(f" Wrote {filename} and '.ok' file. Took {round(time()-start,5)} sec")
 
-class Eventlist:
+class EventlistGeneral:
     def __init__(self) -> None:
         self.dl0_peaks = None
         self.version = None
 
-
     def moving_average(self, x, w):
         return np.convolve(x, np.ones(w), 'valid') / w
-
-
+    
     @staticmethod
     def twos_comp_to_int(val, bits=14):
         if (val & (1 << (bits - 1))) != 0: # if sign bit is set e.g., 8bit: 128-255
@@ -138,68 +136,25 @@ class Eventlist:
             os.makedirs(directory_path)
             print(f"La directory '{directory_path}' è stata creata.")
 
-    def __getPeaks(self, peaks, data):
-        version = data._v_attrs.VERSION
-        if version=='3.1.1' or version=='3.2.0':
-            # If is a DL1
-            start_off = data._v_attrs.wf_start
-            self.dl0_peaks = [pk + start_off for pk in peaks]
-        else:
-            # If is a DL0
-            self.dl0_peaks = peaks
-        return self.dl0_peaks
-    
-    def __getXrange(self, data):
-        version = data._v_attrs.VERSION
-        if version=='3.1.1' or version=='3.2.0':
-            # If is a DL1
-            xr = np.arange(data._v_attrs.wf_start, data._v_attrs.wf_size + data._v_attrs.wf_start)
-        else:
-            # If is a DL0
-            xr = np.arange(len(data))
-        return xr
-
-    def __getWFindex(self, i, data):
-        version = data._v_attrs.VERSION
-        if version=='3.1.1' or version=='3.2.0':
-            # If is a DL1
-            i = int(data._v_attrs.original_wf.split('_')[1])
-        return i
-
-    def __getPKindex(self, j, data):
-        version = data._v_attrs.VERSION
-        if version=='3.1.1' or version=='3.2.0':
-            # If is a DL1
-            if data._v_attrs.n_peaks == 1:
-                return data._v_attrs.peak_idx
-            else:
-                # NOTE: in DL0 wf with more than a PK get a +1
-                return data._v_attrs.peak_idx + 1
-        return j
-
+class EventlistDL0(EventlistGeneral):
     def process_file(self, filename, temperatures, outdir, log = False, startEvent=0, endEvent=-1):
         print("Processing " + filename)
         self.create_directory(outdir)
-
         basename = Path(outdir, os.path.basename(filename))
         outputfilename = f"{Path(basename).with_suffix('.dl2.h5')}"
-        #check if file exists and is not empty
+        # Check if file exists and is not empty
         self.delete_empty_file(outputfilename)
         if os.path.exists(outputfilename):
             return
-
         h5file = open_file(filename, mode="r")
         self.temperatures = temperatures
-
         group = h5file.get_node("/waveforms")
-        
         tstarts = []
         header = f"N_Waveform\tmult\ttstart\tindex_peak\tpeak\tintegral1\tintegral2\tintegral3\thalflife\ttemp"
         f = open(f"{Path(basename).with_suffix('.dl2.txt')}", "w")
         f.write(f"{header}\n")
         dl2_data = []
-
-        #print(header)
+        
         shape_data = -1
         lenwf = -1
         for i, data in tqdm(enumerate(group), total=endEvent+1 if endEvent > 0 else group._g_getnchildren()):
@@ -208,18 +163,16 @@ class Eventlist:
             if endEvent > 0:
                 if i > int(endEvent):
                     break
-            #print(data._v_attrs._f_list("all"))
-            #print(attrs)
-            #tstarts.append(tstart)
+            # Get shape data
             if shape_data < 0:
-                if data._v_attrs.VERSION == "1.1" or data._v_attrs.VERSION == "3.1.1":
+                if data._v_attrs.VERSION == "1.1":
                     shape_data = 1
-                elif data._v_attrs.VERSION == "2.0" or data._v_attrs.VERSION == "3.2.0":
+                elif data._v_attrs.VERSION == "2.0":
                     shape_data = 0
             lenwf = len(data[:,shape_data])
-
+            # Get tstart
             tstart = data._v_attrs.tstart
-
+            # Get max value in wf
             val = np.max(data[:,shape_data])
             if val > 8192:
                 y = np.array(data[:,shape_data].shape)     
@@ -227,37 +180,25 @@ class Eventlist:
                     y[i] = Eventlist.twos_comp_to_int(val)
             else:
                 y = data[:,shape_data]
-
+            # Deep copy of y array 
             arr = y.copy()
+            # Compute moving average
             arrmov = self.moving_average(arr, 15)
-
-            if data._v_attrs.VERSION == "3.1.1" or data._v_attrs.VERSION == "3.2.0":
-                if log == True:
-                    print("recupero mean value and stdev")
-                mmean1 = data._v_attrs.mmean1
-                stdev1 = data._v_attrs.stdev1
-                if not data._v_attrs.isdouble:
-                    peaks  = [data._v_attrs.peak_pos - data._v_attrs.wf_start]
-                else:
-                    peaks, _ = find_peaks(arrmov, height=mmean1 * 2 * 0.9, width=15, distance=25)
-            else:
-                arr3 = arr[:100]
-                mmean1 = arr3.mean()
-                stdev1 = arr3.std()
-                mmean2 = mmean1 * 2 * 0.9 
-                peaks, _ = find_peaks(arrmov, height=mmean2, width=15, distance=25)
-            
+            # Extract peaks
+            arr3 = arr[:100].copy()
+            mmean1 = arr3.mean()
+            stdev1 = arr3.std()
+            mmean2 = mmean1 * 2 * 0.9 
+            peaks, _ = find_peaks(arrmov, height=mmean2, width=15, distance=25)
             deltav = 20
-            # if data._v_attrs.VERSION == "1.1" or data._v_attrs.VERSION == "2.0":
             peaks2 = np.copy(peaks)
-            #filtraggio picchi
+            # Filtering peaks
             for v in peaks2:
                 arrcalcMM = arrmov[v] - arrmov[v-deltav:v+deltav].copy()
-                #80 has been chosen with heuristics on data 
+                # 80 has been chosen with heuristics on data 
                 ind = np.where(arrcalcMM[:] > 80)
-                #remove peaks too small or peaks too close to the end of the wf
+                # Remove peaks too small or peaks too close to the end of the wf
                 if len(ind[0]) == 0 or v > 16000:
-                #if  v > 16000:
                     if log == True:
                         print("delete peak")
                         print(peaks)
@@ -268,10 +209,9 @@ class Eventlist:
                             plt.axvline(x = v, color = 'r') 
                         plt.show()
                     peaks = peaks[peaks != v]
-            
             if log == True:
                 print(f"Waveform num. {i} ##############################")
-                print(f"la waveform num. {i} ha i seguenti peaks: {self.__getPeaks(peaks, data)} e mean value {mmean1} and stdev {stdev1}")
+                print(f"la waveform num. {i} ha i seguenti peaks: {peaks} e mean value {mmean1} and stdev {stdev1}")
 
             if len(peaks) == 0:
                 current_tstart = tstart
@@ -293,7 +233,7 @@ class Eventlist:
                     rowsHalf = [0]
                     try:
                         # Calculation on raw data
-                        arrcalc = arr[v-deltav:]
+                        arrcalc = arr[v-deltav:].copy()
                         blocks_size = 25
                         arrcalc_blocks = np.lib.stride_tricks.sliding_window_view(arrcalc, blocks_size)
                         meanblock  = arrcalc_blocks.mean(axis=1)
@@ -306,8 +246,7 @@ class Eventlist:
                             arrSignal = arrcalc.copy()
                         arrSub = np.subtract(arrSignal, mmean1)
                         integral = np.sum(arrSub)
-                        
-                        # calculation on MM
+                        # Calculation on MM
                         arrcalcMM = arrmov[v-deltav:].copy()
                         arrcalcMM_blocks = np.lib.stride_tricks.sliding_window_view(arrcalcMM, blocks_size)
                         meanblockMM  = arrcalcMM_blocks.mean(axis=1)
@@ -317,17 +256,15 @@ class Eventlist:
                         else:
                             arrSignalMM = arrcalcMM.copy()
                         arrSubMM = np.subtract(arrSignalMM, mmean1)
-                        integralMM = np.sum(arrSubMM)                
-                        
-                        # compare with exponential decay
+                        integralMM = np.sum(arrSubMM)
+                        # Compare with exponential decay
                         arrExp = arrSubMM.copy()
                         rowsHalf=np.where(arrExp[deltav:]<=arrExp[deltav]/2.0)[0]
                         xr = range(deltav, len(arrExp)+deltav)
                         xr2 = range(len(arrExp))
                         valueE = arrExp[deltav] * np.power(2, xr2 * (-1/(rowsHalf[0])))
                         integralExp = np.sum(valueE)
-
-                        #subtract the exponential decay for pileup
+                        # Subtract the exponential decay for pileup
                         if len(peaks) > 1:
                             lenss = v+len(valueE)
                             if lenss > lenwf:
@@ -338,8 +275,217 @@ class Eventlist:
                             #       potrebbero modificarne il valore associato 
                             ss = arrmov[v-deltav:lenss].copy()
                             ss[deltav:lenss] = ss[deltav:lenss] - valueE[0:len(ss[deltav:lenss])]
-                            ss[0:deltav] = np.full(len(ss[0:deltav] ), mmean1)
+                            ss[0:deltav] = np.full(len(ss[0:deltav]), mmean1)
+                        # Plot arr, arrmov and peaks
+                        if log == True:
+                            if j == 0:
+                                plt.figure()
+                                plt.plot(np.arange(len(arr)), arr, color='g', label='arr')
+                                plt.plot(np.arange(len(arr))[:len(arrmov)], arrmov, label='arrmov')
+                                for v in peaks:
+                                    plt.axvline(x = v, color = 'r', label='peaks')
+                                plt.legend()
+                                plt.show()
+                            # Print integrals
+                            print(f"integral {integral}")
+                            print(f"integralMM {integralMM}")
+                            print(f"integralEXP {integralExp}")
+                            # Plot comparison peaks with exponential decay
+                            plt.figure()
+                            plt.plot(range(len(arrSub)),arrSub, label='arrSub')
+                            plt.plot(range(len(arrSubMM)),arrSubMM, label='arrSubMM', color='black')
+                            plt.axvline(x=deltav, label='peak', color = 'r')
+                            plt.plot(xr, valueE, label='exponantial decay', color = 'r')
+                            if len(peaks) > 1:
+                                plt.plot(range(len(ss)),ss-mmean1, label='ss-mmean1')
+                            plt.legend()
+                            plt.show()
+                    except Exception as e:
+                        print(f"EXCEPTION: Peaks non trovati nella waveform {i} del file {filename}")
+                        traceback.print_exception(e)
+                        continue
+                    # Get temperatures
+                    temp = float(self.get_temperature(tstart))
+                    # Wirte on files results
+                    if len(peaks) == 1:
+                        current_tstart = float(tstart)
+                        f.write(f"{i}\t{0}\t{tstart}\t{data[0]}\t{y[peaks[0]]}\t{integral}\t{integralMM}\t{integralExp}\t{rowsHalf[0]}\t{temp:.2f}\n")
+                        dl2_data.append([i, 0, current_tstart, peaks[0], y[peaks[0]], integral, integralMM, integralExp, rowsHalf[0],temp])
+                    else:
+                        current_tstart = float(((peaks[j] - peaks[0]) * 8e-9) + tstart)
+                        f.write(f"{i}\t{j+1}\t{current_tstart}\t{peaks[j]}\t{y[peaks[j]]}\t{integral}\t{integralMM}\t{integralExp}\t{rowsHalf[0]}\t{temp:.2f}\n")
+                        dl2_data.append([i, j+1, current_tstart, peaks[j], y[peaks[j]], integral, integralMM, integralExp, rowsHalf[0],temp])
+                    j = j + 1
+        h5file.close()
+        GFhandler2.write(f"{Path(basename).with_suffix('.dl2.h5')}", dl2_data)
+        f.close()
 
+class EventlistDL1(EventlistGeneral):
+    def __getPeaks(self, peaks, data):
+        start_off = data._v_attrs.wf_start
+        self.dl0_peaks = [pk + start_off for pk in peaks]
+        return self.dl0_peaks
+    
+    def __getXrange(self, data):
+        return np.arange(data._v_attrs.wf_start, data._v_attrs.wf_size + data._v_attrs.wf_start)
+
+    def __getWFindex(self, data):
+        return int(data._v_attrs.original_wf.split('_')[1])
+
+    def __getPKindex(self, j, data):
+        if data._v_attrs.n_peaks == 1:
+            return 0
+        else:
+            # NOTE: in DL0 wf with more than a PK get a +1
+            return data._v_attrs.peak_idx + j + 1
+
+    def process_file(self, filename, temperatures, outdir, log = False, startEvent=0, endEvent=-1):
+        print("Processing " + filename)
+        self.create_directory(outdir)
+        basename = Path(outdir, os.path.basename(filename))
+        outputfilename = f"{Path(basename).with_suffix('.dl2.h5')}"
+        # Check if file exists and is not empty
+        self.delete_empty_file(outputfilename)
+        if os.path.exists(outputfilename):
+            return
+        # Open h5file
+        h5file = open_file(filename, mode="r")
+        self.temperatures = temperatures
+        group = h5file.get_node("/waveforms")
+        tstarts = []
+        header = f"N_Waveform\tmult\ttstart\tindex_peak\tpeak\tintegral1\tintegral2\tintegral3\thalflife\ttemp"
+        f = open(f"{Path(basename).with_suffix('.dl2.txt')}", "w")
+        f.write(f"{header}\n")
+        dl2_data = []
+
+        shape_data = -1
+        lenwf = -1
+        for i, data in tqdm(enumerate(group), total=endEvent+1 if endEvent > 0 else group._g_getnchildren()):
+            if i < int(startEvent):
+                continue
+            if endEvent > 0:
+                if i > int(endEvent):
+                    break
+            if shape_data < 0:
+                if data._v_attrs.VERSION == "3.1.1":
+                    shape_data = 1
+                elif data._v_attrs.VERSION == "3.2.0":
+                    shape_data = 0
+            lenwf = len(data[:,shape_data])
+            # Get tstart
+            tstart = data._v_attrs.tstart
+            # Get maximum value of wf
+            val = np.max(data[:,shape_data])
+            if val > 8192:
+                y = np.array(data[:,shape_data].shape)     
+                for i, val in enumerate(data[:,shape_data]):
+                    y[i] = Eventlist.twos_comp_to_int(val)
+            else:
+                y = data[:,shape_data]
+            # Deep copy of y array 
+            arr = y.copy()
+            # Compute moving average
+            arrmov = self.moving_average(arr, 15)
+            # Get metadata of wf and its peaks
+            if log == True:
+                print("recupero mean value and stdev")
+            mmean1 = data._v_attrs.mmean1
+            stdev1 = data._v_attrs.stdev1
+            if data._v_attrs.n_peaks == 0:
+                # If in DL1 we didn't find any peak the list should be empty
+                peaks = []
+            elif not data._v_attrs.isdouble:
+                # If it is a single event event 
+                peaks  = [data._v_attrs.peak_pos - data._v_attrs.wf_start]
+            else:
+                # If it is a double event
+                peaks, _ = find_peaks(arrmov, height=mmean1*2* 0.9, width=15, distance=25)
+            deltav = 20
+            peaks2 = np.copy(peaks)
+            for v in peaks2:
+                arrcalcMM = arrmov[v] - arrmov[v-deltav:v+deltav].copy()
+                # 80 has been chosen with heuristics on data 
+                ind = np.where(arrcalcMM[:] > 80)
+                # Remove peaks too small or peaks too close to the end of the wf
+                if len(ind[0]) == 0 or v > 16000:
+                    if log == True:
+                        print("delete peak")
+                        print(peaks)
+                        plt.figure()
+                        plt.plot(range(len(arr)),arr, color='g')
+                        plt.plot(range(len(arrmov)),arrmov)
+                        for v in peaks:
+                            plt.axvline(x = v, color = 'r') 
+                        plt.show()
+                    peaks = peaks[peaks != v]
+            if log == True:
+                print(f"Waveform num. {i} ##############################")
+                print(f"la waveform num. {i} ha i seguenti peaks: {self.__getPeaks(peaks, data)} e mean value {mmean1} and stdev {stdev1}")
+
+            if len(peaks) == 0:
+                current_tstart = tstart
+                temp = self.get_temperature(tstart)
+                f.write(f"{self.__getWFindex(data)}\t{0}\t{tstart}\t{-1}\t{-1}\t{-1}\t{-1}\t{-1}\t{-1}\t{temp:.2f}\n")
+                dl2_data.append([self.__getWFindex(data), 0, tstart, -1, -1, -1, -1, -1, -1, temp])
+                if log == True:
+                    print(f"{i}\tEMPTY")
+                    plt.figure()
+                    plt.plot(range(len(arr)),arr, color='g')
+                    plt.plot(range(len(arrmov)),arrmov)
+                    plt.show()
+            else:
+                j=0
+                for v in peaks:
+                    integral = 0
+                    integralMM = 0
+                    integralExp = 0
+                    rowsHalf = [0]
+                    try:
+                        # Calculation on raw data
+                        arrcalc = arr[v-deltav:].copy()
+                        blocks_size = 25
+                        arrcalc_blocks = np.lib.stride_tricks.sliding_window_view(arrcalc, blocks_size)
+                        meanblock  = arrcalc_blocks.mean(axis=1)
+                        rowsL = np.where(np.logical_and(meanblock > mmean1 - stdev1, meanblock < mmean1 + stdev1))[0]
+                        if len(rowsL) > 0:
+                            # Since sliding windows have stride 1, mean_block indices are 0:(len(arr)-block_size)
+                            # So to index arr_calc you can use rowsL[0] directly
+                            arrSignal = arrcalc[:rowsL[0]].copy()
+                        else:
+                            arrSignal = arrcalc.copy()
+                        arrSub = np.subtract(arrSignal, mmean1)
+                        integral = np.sum(arrSub)
+                        # Calculation on MM
+                        arrcalcMM = arrmov[v-deltav:].copy()
+                        arrcalcMM_blocks = np.lib.stride_tricks.sliding_window_view(arrcalcMM, blocks_size)
+                        meanblockMM  = arrcalcMM_blocks.mean(axis=1)
+                        rowsLMM = np.where(np.logical_and(meanblockMM > mmean1 - stdev1, meanblockMM < mmean1 + stdev1))[0]
+                        if len(rowsLMM) > 0:
+                            arrSignalMM = arrcalcMM[:rowsLMM[0]].copy()
+                        else:
+                            arrSignalMM = arrcalcMM.copy()
+                        arrSubMM = np.subtract(arrSignalMM, mmean1)
+                        integralMM = np.sum(arrSubMM)                
+                        # Compare with exponential decay
+                        arrExp = arrSubMM.copy()
+                        rowsHalf=np.where(arrExp[deltav:]<=arrExp[deltav]/2.0)[0]
+                        xr = range(deltav, len(arrExp)+deltav)
+                        xr2 = range(len(arrExp))
+                        valueE = arrExp[deltav] * np.power(2, xr2 * (-1/(rowsHalf[0])))
+                        integralExp = np.sum(valueE)
+                        # Subtract the exponential decay for pileup
+                        if len(peaks) > 1:
+                            lenss = v+len(valueE)
+                            if lenss > lenwf:
+                                lenss = lenwf
+                            # NOTE: togliendo .copy() arrmov cambia. Con lui cambiano anche i plots e i valori degli integrali.
+                            #       in teoria non dovrebbe cambiare nulla, ma ChatGPT suggerisce che talvolta quando due array 
+                            #       condividono lo stesso spazio in memoria alcune operazioni anche se non eseguono accessi diretti
+                            #       potrebbero modificarne il valore associato 
+                            ss = arrmov[v-deltav:lenss].copy()
+                            ss[deltav:lenss] = ss[deltav:lenss] - valueE[0:len(ss[deltav:lenss])]
+                            ss[0:deltav] = np.full(len(ss[0:deltav]), mmean1)
+                        # Plot arr, arrmov and peaks
                         if log == True:
                             if j == 0:
                                 plt.figure()
@@ -350,11 +496,11 @@ class Eventlist:
                                     plt.axvline(x = v, color = 'r', label='peaks')
                                 plt.legend()
                                 plt.show()
-                            
+                            # Print integrals
                             print(f"integral {integral}")
                             print(f"integralMM {integralMM}")
                             print(f"integralEXP {integralExp}")
-
+                            # Plot comparison peaks with exponential decay
                             plt.figure()
                             plt.plot(range(len(arrSub)),arrSub, label='arrSub')
                             plt.plot(range(len(arrSubMM)),arrSubMM, label='arrSubMM', color='black')
@@ -364,33 +510,53 @@ class Eventlist:
                                 plt.plot(range(len(ss)),ss-mmean1, label='ss-mmean1')
                             plt.legend()
                             plt.show()
-
                     except Exception as e:
                         print(f"EXCEPTION: Peaks non trovati nella waveform {i} del file {filename}")
                         traceback.print_exception(e)
                         continue
-
+                    # Get temperatures
                     temp = float(self.get_temperature(tstart))
-
-                    if len(peaks) == 1:
-                        current_tstart = float(tstart)
-                        # f.write(f"{i}\t{0}\t{tstart}\t{data[0]}\t{y[peaks[0]]}\t{integral}\t{integralMM}\t{integralExp}\t{rowsHalf[0]}\t{temp:.2f}\n")
-                        f.write(f"{self.__getWFindex(i, data)}\t{self.__getPKindex(0, data)}\t{tstart}\t{self.__getPeaks(peaks, data)[0]}\t{y[peaks[0]]}\t{integral}\t{integralMM}\t{integralExp}\t{rowsHalf[0]}\t{temp:.2f}\n")
-                        # dl2_data.append([i, 0, current_tstart, peaks[0], y[peaks[0]], integral, integralMM, integralExp, rowsHalf[0],temp])
-                        dl2_data.append([self.__getWFindex(i, data), self.__getPKindex(0, data), current_tstart, self.__getPeaks(peaks, data)[0], y[peaks[0]], integral, integralMM, integralExp, rowsHalf[0],temp])
-                    else:
-                        current_tstart = float(((peaks[j] - peaks[0]) * 8e-9) + tstart)
-                        # f.write(f"{i}\t{j+1}\t{current_tstart}\t{peaks[j]}\t{y[peaks[j]]}\t{integral}\t{integralMM}\t{integralExp}\t{rowsHalf[0]}\t{temp:.2f}\n")
-                        f.write(f"{self.__getWFindex(i, data)}\t{self.__getPKindex(j+1, data)}\t{current_tstart}\t{self.__getPeaks(peaks, data)[j]}\t{y[peaks[j]]}\t{integral}\t{integralMM}\t{integralExp}\t{rowsHalf[0]}\t{temp:.2f}\n")
-                        # dl2_data.append([i, j+1, current_tstart, peaks[j], y[peaks[j]], integral, integralMM, integralExp, rowsHalf[0],temp])
-                        dl2_data.append([self.__getWFindex(i, data), self.__getPKindex(j+1, data), current_tstart, self.__getPeaks(peaks, data)[j], y[peaks[j]], integral, integralMM, integralExp, rowsHalf[0], temp])
-
+                    # Wirte on files results
+                    original_tstart = data._v_attrs.pk0_tstart
+                    original_pk0    = data._v_attrs.pk0_pos
+                    current_tstart = ((self.__getPeaks(peaks, data)[j] - original_pk0) * 8e-9) + original_tstart
+                    f.write(f"{self.__getWFindex(data)}\t{self.__getPKindex(j, data)}\t{current_tstart}\t{self.__getPeaks(peaks, data)[j]}\t{y[peaks[j]]}\t{integral}\t{integralMM}\t{integralExp}\t{rowsHalf[0]}\t{temp:.2f}\n")
+                    dl2_data.append([self.__getWFindex(data), self.__getPKindex(j, data), current_tstart, self.__getPeaks(peaks, data)[j], y[peaks[j]], integral, integralMM, integralExp, rowsHalf[0], temp])
                     j = j + 1
-
         h5file.close()
         GFhandler2.write(f"{Path(basename).with_suffix('.dl2.h5')}", dl2_data)
-        
         f.close()
+
+
+class Eventlist:
+    def __init__(self, from_dl1=False) -> None:
+        self.from_dl1=from_dl1
+        if not self.from_dl1:
+            self.evntlist = EventlistDL0()
+        else:
+            self.evntlist = EventlistDL1()
+
+    def moving_average(self, x, w):
+        self.evntlist.moving_average(x, w)
+    
+    def twos_comp_to_int(self, val, bits=14):
+        self.twos_comp_to_int(val, bits=bits)
+
+    def process_temps_file(self, filename):
+        self.evntlist.process_temps_file(filename)
+    
+    def get_temperature(self, tstart):
+        self.evntlist.get_temperature(tstart)
+
+    def delete_empty_file(self, nome_file):
+        self.evntlist.delete_empty_file(nome_file)
+
+    def create_directory(self, directory_path):
+        self.evntlist.create_directory(directory_path)
+
+    def process_file(self, filename, temperatures, outdir, log=False, startEvent=0, endEvent=-1):
+        self.evntlist.process_file(filename, temperatures, outdir, log=log, startEvent=startEvent, endEvent=endEvent)
+
 
 if __name__ == '__main__':
